:hammer: Move grab_files from cli to io
Refactoring of grab_files and it's tests from cli to io.
